requirements.txt
anyio==3.6.1
certifi==2022.6.15
click==8.1.3
cloudinary==1.29.0
dnspython==2.2.1
fastapi==0.81.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
motor==3.0.0
pydantic==1.10.0
pymongo==4.2.0
python-dateutil==2.8.2
python-dotenv==0.20.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing-extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
pycryptodome==3.15.0
